feat: new admin middleware
