+++ app/Middleware/AuthenticateMiddleware.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace App\Middleware;
+
+use Core\Http\Request;
+use Core\Http\Middleware\Middleware;
+use Lib\Authentication\Auth;
+use Lib\FlashMessage;
+
+class AuthenticateMiddleware implements Middleware
+{
+    public function handle(Request $request): void
+    {
+        if (!Auth::check()) {
+            FlashMessage::danger('Você deve estar logado para acessar essa página');
+            $this->redirectTo(route('users.login'));
+        }
+    }
+
+    private function redirectTo(string $location): void
+    {
+        header('Location: ' . $location);
+        exit;
+    }
+}
