+++ config/App.php
@@ -5,6 +5,7 @@ namespace Config;
 class App
 {
     public static array $middlewareAliases = [
-        'auth' => \App\Middleware\Authenticate::class
+        'auth' => \App\Middleware\AuthenticateMiddleware::class,
+        'admin' => \App\Middleware\AdminMiddleware::class
     ];
 }
