+++ config/routes.php
@@ -14,9 +14,11 @@ Route::post('/login', [AuthenticationsController::class, 'authenticate'])->name(
 
 Route::middleware('auth')->group(function () {
     Route::get('/', [HomeController::class, 'index'])->name('root');
-    Route::get('/admin', [AdminController::class, 'index'])->name('admin.index');
     Route::get('/mushrooms', [MushroomController::class, 'index'])->name('mushrooms.index');
 
-    // Logout
     Route::get('/logout', [AuthenticationsController::class, 'destroy'])->name('users.logout');
+});
+
+Route::middleware('admin')->group(function () {
+    Route::get('/admin', [AdminController::class, 'index'])->name('admin.index');
 });
