+++ core/Router/Route.php
@@ -19,8 +19,7 @@ class Route
         private string $uri,
         private string $controllerName,
         private string $actionName
-    ) {
-    }
+    ) {}
 
     public function name(string $name): void
     {
